+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for Open Measurement SDK JavaScript clients
 
+## 1.6.6 - 2026-05-18
+- Improved precision handling in compliance and validation event data.
+- CTV reference app: Fix scrolling viewability geometry.
+- Export VideoPlayerState, InteractionType, CreativeType, ErrorType, and ImpressionType to OmidSessionClient. 
+
 ## 1.6.5 - 2026-04-01
 - Update version to match iOS SDK; no changes from 1.6.4